data_analysis_agent/pyproject.toml
[project]
name = "data-analysis-agent"
version = "1.0.0"
description = "基于大语言模型的智能数据分析Agent"
authors = [
    {name = "Data Analysis Team"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "pandas>=2.0.0", 
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "jinja2>=3.1.0",
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.9.0",
    "openpyxl>=3.1.0",
    "python-docx>=0.8.11",
    "seaborn>=0.12.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
]

[project.scripts]
data-agent = "data_analysis_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []